+++ docs/index.html
@@ -1,19 +1,32 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <title>React Vapor</title>
-
     <meta charset="UTF-8">
+    <script src="https://code.jquery.com/jquery-3.3.1.js" integrity="sha256-2Kok7MbOyxpgUVvAk/HJ2jigOSYS2auK4Pfzbm7uH60=" crossorigin="anonymous"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.38.0/codemirror.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/redux/3.7.2/redux.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/react/15.6.2/react.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/react-dom/15.6.2/react-dom.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/react-redux/5.0.6/react-redux.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/react-bootstrap/0.31.3/react-bootstrap.min.js"></script>
+    <script src="https://static.cloud.coveo.com/styleguide/v2.19.1/js/VaporSVG.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/underscore.js/1.9.1/underscore.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.22.2/moment.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/rc-slider@8.2.3/lib/index.min.js"></script>
 </head>
+
 <body class="coveo-styleguide">
-<div class="header flex" style="min-height: 65px;">
-    <div class="header-section h1 p2 flex flex-center">
-        React Vapor
+    <div class="header flex" style="min-height: 65px;">
+        <div class="header-section h1 p2 flex flex-center">
+            React Vapor
+        </div>
     </div>
-</div>
-<div class="container p2">
-    <div id="App"></div>
-</div>
-<script src="assets/bundle.js"></script>
+    <div class="container p2">
+        <div id="App"></div>
+    </div>
+    <script defer src="assets/bundle.js"></script>
 </body>
+
 </html>
