+++ webpack.config.js
@@ -1,27 +1,58 @@
-const webpack = require('webpack');
 const path = require('path');
-const isTravis = !!process.env.TRAVIS;
+const webpack = require('webpack');
+const isTravis = process.env.TRAVIS;
+const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
 
-/**
- * Config file for the documentation project
- */
-module.exports = {
-    entry: './docs/Index.tsx',
-    mode: 'development',
+const isDocs = !!process.env.DOCS;
+let contextDependentConfig;
+if (isDocs) {
+    contextDependentConfig = {
+        mode: 'development',
+        entry: './docs/Index.tsx',
+        output: {
+            path: path.join(__dirname, '/docs/assets'),
+            publicPath: '/assets/',
+            filename: 'bundle.js',
+        },
+        devtool: 'source-map',
+        devServer: {
+            contentBase: './docs',
+        },
+        plugins: [
+            new webpack.DefinePlugin({
+                WEBPACK_DEFINED_VERSION: JSON.stringify(require('./package.json').version),
+            }),
+        ],
+    }
+} else {
+   contextDependentConfig = {
+    entry: './Index.ts',
+    mode: 'production',
+    optimization: {minimize: false},
     output: {
-        path: path.join(__dirname, '/docs/assets'),
-        publicPath: '/assets/',
-        filename: 'bundle.js',
-    },
-    devtool: 'source-map',
-    resolve: {
-        extensions: ['.ts', '.tsx', '.js', '.jsx'],
+        path: path.join(__dirname, '/dist'),
+        filename: 'react-vapor.js',
+        library: ['ReactVapor'],
+        libraryTarget: 'umd',
     },
     plugins: [
         new webpack.DefinePlugin({
             WEBPACK_DEFINED_VERSION: JSON.stringify(require('./package.json').version),
+            'process.env.NODE_ENV': JSON.stringify('production'),
         }),
+        // new BundleAnalyzerPlugin(), // Uncomment to analyze the bundle
     ],
+   } 
+}
+
+/**
+ * Config file for the packaged library
+ */
+const config = {
+    ...contextDependentConfig,
+    resolve: {
+        extensions: ['.ts', '.tsx', '.js'],
+    },
     module: {
         rules: [
             {
@@ -33,22 +64,17 @@ module.exports = {
                     options: {
                         configFile: './node_modules/tsjs/tslint.json',
                         tsConfigFile: './tsconfig.json',
-                        emitErrors: isTravis,
+                        emitErrors: true,
                         failOnHint: isTravis,
                     },
                 },
             },
             {
-                test: /\.(ts|tsx)$/,
+                test: /\.ts(x?)$/,
                 loader: 'ts-loader',
                 options: {
                     compiler: 'ttypescript',
                 },
-            },
-            {
-                test: /\.scss$/,
-                exclude: [/src\/components/],
-                loader: 'style-loader!css-loader!postcss-loader!sass-loader',
             },
             {
                 test: /\.css$/,
@@ -83,6 +109,12 @@ module.exports = {
                     },
                 ],
             },
+            // Provide jQuery=require('jquery') to use the same jquery instance.
+            // See http://reactkungfu.com/2015/10/integrating-jquery-chosen-with-webpack-using-imports-loader/ for more infos.
+            {
+                test: require.resolve('chosen-js'),
+                loader: 'imports-loader?jQuery=jquery',
+            },
             {
                 test: /\.png$/,
                 loader: 'file-loader?mimetype=image/png',
@@ -91,16 +123,21 @@ module.exports = {
                 test: /\.(ttf|eot|woff|svg)(\?v=[0-9]\.[0-9]\.[0-9])?$/,
                 loader: 'file-loader',
             },
-
-            // Provide jQuery=require('jquery') to use the same jquery instance.
-            // See http://reactkungfu.com/2015/10/integrating-jquery-chosen-with-webpack-using-imports-loader/ for more infos.
-            {
-                test: require.resolve('chosen-js'),
-                loader: 'imports-loader?jQuery=jquery',
-            },
         ],
     },
-    devServer: {
-        contentBase: './docs',
+    externals: {
+        codemirror: 'CodeMirror',
+        jquery: '$',
+        react: 'React',
+        'react-dom': 'ReactDOM',
+        'react-bootstrap': 'ReactBootstrap',
+        'react-redux': 'ReactRedux',
+        'rc-slider': 'Slider',
+        'coveo-styleguide': 'VaporSVG',
+        redux: 'Redux',
+        underscore: '_',
+        moment: 'moment',
     },
 };
+
+module.exports = config;
