--- webpack.config.prod.minify.js
@@ -1,6 +1,0 @@
-const prodConfig = require('./webpack.config.prod');
-
-prodConfig.output.filename = 'react-vapor.min.js';
-prodConfig.optimization = {minimize: true};
-
-module.exports = prodConfig;
